Implement ImgLoader component
A pattern to load an image component with a placeholder.

Includes:

- an Anyloader component that can be used to create a wrapper for any
  specific image component.

Specific wrappers implemented:

- ImgLoader to load a html img component
- ImageLoader to load a Semantic UI Image component

In both cases the `placeholder` property must be specified to define
the placeholder content. This is typically an svg or any other html
markup that can be displayed until the image is loaded.

Similarly, a lightweight wrapper can be created to support any specific
image component by using the AnyLoader component.