Add createMissingScaleSrc option as fallback until scales defined #277
The need emerges because scales is not available until the scales are
defined, for example in the add view of a new image. This function
makes it possible to use a fallback redirect view such as
`.../@@images/image/SCALENAME` which won't be cacheable but it will
appear even if the src points to the image instance and not an actual
image url.