Fix image loading with and without blurhash, fixes jump #276
Two cases are fixed:

1. Without blurhash

Fix edge case when there is no placeholder (or blurhash). In this case
shortcut to show the image component directly without any loading.

This fixes a "jump" during image loading for components that don't
define a blurhash (and neither a placeholder).

2. With blurhash

Fix the checking of the aspect ratio correctly, this is important to
allow the component to work with or without aspect ratio. The height
is controlled only if there is no aspect ratio set.