Image loader: fix blurhash canvas aspect ratio #276
This problem is apparent with `aspect-ratio: cover`, while it was also
present with `aspect-ratio: fill`, it's not easily observable. The
solution is to make the blurhash canvas have the same aspect ratio as
the *original* image (not the rendered styles), this way aspect-ratio
will work identically to the to-be-loaded image.

Also deprecate prop `resolutionY` which is now calculated from the
ratio.