Fix blurhash zero width resizing bug (#143)
* Fix blurhash zero width resizing bug

Under some circumstances the ResizeObserver is called on a zero offset
image, which would cause the height to be reset and the blurhash
disappear.

This might be a ResizeObserver bug in newer Chrome, or subject of a
race condition. Precise reason unconfirmed.

* Fix ignoring blurhash width if width (or size) was a string

A number is executed correctly as a style, but a string is silently
ignored unless it has the `px` measure in the end. Fix this when we
apply the `width, height` style attributes to the blurhash canvas.

* Update fast-blurhash with recent blurhash fixes

The fast-blurhash file is in plain js and it must be kept in sync with
the react components.

* update mrs-developer

Co-authored-by: Balázs Reé <ree@greenfinity.hu>
Co-authored-by: David Glick <david@glicksoftware.com>